Separate action generation part as a new class, TSActionGenerator.
