+++ org.lflang/src/org/lflang/generator/ts/TSReactorGenerator.kt
@@ -72,21 +72,6 @@ class TSReactorGenerator(
     }
 
     /**
-     * Return a TS type for the specified action.
-     * If the type has not been specified, return
-     * "Present" which is the base type for Actions.
-     * @param action The action
-     * @return The TS type.
-     */
-    private fun getActionType(action: Action): String {
-        if (action.type != null) {
-            return action.type.getTargetType()
-        } else {
-            return "Present"
-        }
-    }
-
-    /**
      * Return a TS type for the specified port.
      * If the type has not been specified, return
      * "Present" which is the base type for ports.
@@ -127,35 +112,6 @@ class TSReactorGenerator(
             ${" |    "..arguments.joinToString(", \n")}
                 |)
             """.trimMargin()}
-    }
-
-    private fun generateActions(reactor: Reactor): String {
-        val actionInstantiations = LinkedList<String>()
-        for (action in reactor.actions) {
-            // Shutdown actions are handled internally by the
-            // TypeScript reactor framework. There would be a
-            // duplicate action if we included the one generated
-            // by LF.
-            if (action.name != "shutdown") {
-                var actionArgs = "this, __Origin." + action.origin
-                if (action.minDelay != null) {
-                    // Actions in the TypeScript target are constructed
-                    // with an optional minDelay argument which defaults to 0.
-                    if (action.minDelay.parameter != null) {
-                        actionArgs+= ", " + action.minDelay.parameter.name
-                    } else {
-                        actionArgs+= ", " + action.minDelay.getTargetValue()
-                    }
-                }
-                actionInstantiations.add(
-                    "this.${action.name} = new __Action<${getActionType(action)}>($actionArgs);")
-            }
-        }
-        return with(PrependOperator) {
-            """
-            ${" |"..actionInstantiations.joinToString("\n")}
-            """.trimMargin()
-        }
     }
 
     // TODO(hokeun): Split this method into smaller methods.
@@ -258,22 +214,12 @@ class TSReactorGenerator(
         }
 
         val stateGenerator = TSStateGenerator(tsGenerator, reactor)
-
         pr(stateGenerator.generateClassProperties())
         pr(reactorConstructor, stateGenerator.generateInstantiations())
 
-        // Next handle actions.
-        for (action in reactor.actions) {
-            // Shutdown actions are handled internally by the
-            // TypeScript reactor framework. There would be a
-            // duplicate action if we included the one generated
-            // by LF.
-            if (action.name != "shutdown") {
-                pr(action.name + ": __Action<" + getActionType(action) + ">;")
-            }
-        }
-
-        pr(reactorConstructor, generateActions(reactor))
+        val actionGenerator = TSActionGenerator(tsGenerator, reactor)
+        pr(reactorConstructor, actionGenerator.generateClassProperties())
+        pr(reactorConstructor, actionGenerator.generateInstantiations())
 
         // Next handle inputs.
         for (input in reactor.inputs) {
